requirements_test.txt
-e .
pytest==3.0.2
pytest-cov==2.3.1
pytest-django==3.0.0
flake8==3.0.4
codecov==2.0.9
twine==1.9.1
wheel==0.30.0
raven==5.19.0
directory-sso-api-client
directory-api-client
directory-forms-api-client
